change to long

+++ tools/SqlSharp/SqlSharpLit/Common/ParserLit/SqlIdentity.cs
@@ -3,7 +3,7 @@ namespace SqlSharpLit.Common.ParserLit;
 public class SqlIdentity
 {
     public static SqlIdentity Default => new();
-    public int Seed { get; set; }
+    public long Seed { get; set; }
     public int Increment { get; set; }
 
     public string ToSql()
